div fix

+++ programming_paradigm/robust_division_calculator.py
@@ -1,6 +1,6 @@
 def safe_divide(numerator, denominator):
     try:
-        return int(numerator) / int(denominator)
+        return float(numerator) / float(denominator)
     except ZeroDivisionError:
         return "You can't divide by zero!"
     except ValueError:
